XMLHttpRequest/abort-during-open.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: abort() during OPEN</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#the-abort()-method" data-tested-assertations="/following-sibling::ol/li[6] /following-sibling::ol/li[7]" />
  </head>
  <body>
    <div id="log"></div>
    <script>
      var test = async_test()
      test.step(function() {
        var client = new XMLHttpRequest()
        client.open("GET", "...")
        client.onreadystatechange = function() {
          test.step(function() {
            assert_unreached()
          })
        }
        client.abort();
        assert_equals(client.readyState, 0)
      })
      test.done()
    </script>
  </body>
</html>
